modifications from PR review
