+++ packages/suite-base/src/components/AppSettingsDialog/AppSettingsDialog.style.ts
@@ -1,0 +1,83 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+import { makeStyles } from "tss-react/mui";
+
+export const useStyles = makeStyles()((theme) => ({
+  layoutGrid: {
+    display: "grid",
+    gap: theme.spacing(2),
+    height: "70vh",
+    paddingLeft: theme.spacing(1),
+    overflowY: "hidden",
+    [theme.breakpoints.up("sm")]: {
+      gridTemplateColumns: "auto minmax(0, 1fr)",
+    },
+  },
+  logo: {
+    width: 212,
+    height: "auto",
+    marginLeft: theme.spacing(-1),
+  },
+  tabPanel: {
+    display: "none",
+    marginRight: "-100%",
+    width: "100%",
+    padding: theme.spacing(0, 4, 4),
+  },
+  tabPanelActive: {
+    display: "block",
+  },
+  checkbox: {
+    "&.MuiCheckbox-root": {
+      paddingTop: 0,
+    },
+  },
+  dialogActions: {
+    position: "sticky",
+    backgroundColor: theme.palette.background.paper,
+    borderTop: `${theme.palette.divider} 1px solid`,
+    padding: theme.spacing(1),
+    bottom: 0,
+  },
+  formControlLabel: {
+    "&.MuiFormControlLabel-root": {
+      alignItems: "start",
+    },
+  },
+  tab: {
+    svg: {
+      fontSize: "inherit",
+    },
+    "> span, > .MuiSvgIcon-root": {
+      display: "flex",
+      color: theme.palette.primary.main,
+      marginRight: theme.spacing(1.5),
+      height: theme.typography.pxToRem(21),
+      width: theme.typography.pxToRem(21),
+    },
+    [theme.breakpoints.up("sm")]: {
+      textAlign: "right",
+      flexDirection: "row",
+      justifyContent: "flex-start",
+      alignItems: "center",
+      minHeight: "auto",
+      paddingTop: theme.spacing(1.5),
+      paddingBottom: theme.spacing(1.5),
+    },
+  },
+  indicator: {
+    [theme.breakpoints.up("sm")]: {
+      right: 0,
+      width: "100%",
+      backgroundColor: theme.palette.action.hover,
+      borderRadius: theme.shape.borderRadius,
+    },
+  },
+  dialogTitle: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    fontSize: theme.typography.h3.fontSize,
+  },
+}));
