+++ packages/suite-base/src/components/AppSettingsDialog/AppSettingsDialog.tsx
@@ -26,15 +26,16 @@ import {
 } from "@mui/material";
 import { MouseEvent, SyntheticEvent, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { makeStyles } from "tss-react/mui";
 
 import { AppSetting } from "@lichtblick/suite-base/AppSetting";
+import { useStyles } from "@lichtblick/suite-base/components/AppSettingsDialog/AppSettingsDialog.style";
+import { APP_SETTINGS_ABOUT_ITEMS } from "@lichtblick/suite-base/components/AppSettingsDialog/contants";
+import { AppSettingsTab } from "@lichtblick/suite-base/components/AppSettingsDialog/types";
 import CopyButton from "@lichtblick/suite-base/components/CopyButton";
 import { ExperimentalFeatureSettings } from "@lichtblick/suite-base/components/ExperimentalFeatureSettings";
 import ExtensionsSettings from "@lichtblick/suite-base/components/ExtensionsSettings";
 import LichtblickLogoText from "@lichtblick/suite-base/components/LichtblickLogoText";
 import Stack from "@lichtblick/suite-base/components/Stack";
-import { LICHTBLICK_DOCUMENTATION_LINK } from "@lichtblick/suite-base/constants/documentation";
 import { useAppContext } from "@lichtblick/suite-base/context/AppContext";
 import {
   useWorkspaceStore,
@@ -53,127 +54,6 @@ import {
   TimeFormat,
   TimezoneSettings,
 } from "./settings";
-
-const useStyles = makeStyles()((theme) => ({
-  layoutGrid: {
-    display: "grid",
-    gap: theme.spacing(2),
-    height: "70vh",
-    paddingLeft: theme.spacing(1),
-    overflowY: "hidden",
-    [theme.breakpoints.up("sm")]: {
-      gridTemplateColumns: "auto minmax(0, 1fr)",
-    },
-  },
-  logo: {
-    width: 212,
-    height: "auto",
-    marginLeft: theme.spacing(-1),
-  },
-  tabPanel: {
-    display: "none",
-    marginRight: "-100%",
-    width: "100%",
-    padding: theme.spacing(0, 4, 4),
-  },
-  tabPanelActive: {
-    display: "block",
-  },
-  checkbox: {
-    "&.MuiCheckbox-root": {
-      paddingTop: 0,
-    },
-  },
-  dialogActions: {
-    position: "sticky",
-    backgroundColor: theme.palette.background.paper,
-    borderTop: `${theme.palette.divider} 1px solid`,
-    padding: theme.spacing(1),
-    bottom: 0,
-  },
-  formControlLabel: {
-    "&.MuiFormControlLabel-root": {
-      alignItems: "start",
-    },
-  },
-  tab: {
-    svg: {
-      fontSize: "inherit",
-    },
-    "> span, > .MuiSvgIcon-root": {
-      display: "flex",
-      color: theme.palette.primary.main,
-      marginRight: theme.spacing(1.5),
-      height: theme.typography.pxToRem(21),
-      width: theme.typography.pxToRem(21),
-    },
-    [theme.breakpoints.up("sm")]: {
-      textAlign: "right",
-      flexDirection: "row",
-      justifyContent: "flex-start",
-      alignItems: "center",
-      minHeight: "auto",
-      paddingTop: theme.spacing(1.5),
-      paddingBottom: theme.spacing(1.5),
-    },
-  },
-  indicator: {
-    [theme.breakpoints.up("sm")]: {
-      right: 0,
-      width: "100%",
-      backgroundColor: theme.palette.action.hover,
-      borderRadius: theme.shape.borderRadius,
-    },
-  },
-  dialogTitle: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    fontSize: theme.typography.h3.fontSize,
-  },
-}));
-
-type SectionKey = "documentation" | "legal";
-
-const aboutItems = new Map<
-  SectionKey,
-  {
-    subheader: string;
-    links: { title: string; url?: string }[];
-  }
->([
-  [
-    "documentation",
-    {
-      subheader: "Documentation",
-      links: [
-        {
-          title: "Check out our documentation",
-          url: LICHTBLICK_DOCUMENTATION_LINK,
-        },
-      ],
-    },
-  ],
-  [
-    "legal",
-    {
-      subheader: "Legal",
-      links: [
-        {
-          title: "License terms",
-          url: "https://github.com/lichtblick-suite/lichtblick/blob/main/LICENSE",
-        },
-      ],
-    },
-  ],
-]);
-
-export type AppSettingsTab =
-  | "general"
-  | "privacy"
-  | "extensions"
-  | "experimental-features"
-  | "about";
 
 const selectWorkspaceInitialActiveTab = (store: WorkspaceContextStore) =>
   store.dialogs.preferences.initialTab;
@@ -307,11 +187,11 @@ export function AppSettingsDialog(
                   getText={() => LICHTBLICK_SUITE_VERSION?.toString() ?? ""}
                 />
               </Stack>
-              {[aboutItems.get("documentation"), aboutItems.get("legal")].map((item) => {
+              {Array.from(APP_SETTINGS_ABOUT_ITEMS.values()).map((item) => {
                 return (
-                  <Stack key={item?.subheader} gap={1}>
-                    {item?.subheader && <Typography>{item.subheader}</Typography>}
-                    {item?.links.map((link) => (
+                  <Stack key={item.subheader} gap={1}>
+                    {item.subheader && <Typography>{item.subheader}</Typography>}
+                    {item.links.map((link) => (
                       <Link
                         variant="body2"
                         underline="hover"
