+++ packages/suite-base/src/components/AppSettingsDialog/types.ts
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+export type AppSettingsSectionKey = "documentation" | "legal";
+
+export type AppSettingsTab =
+  | "general"
+  | "privacy"
+  | "extensions"
+  | "experimental-features"
+  | "about";
